refactor(connlib): move error metric to the event-loops
Record the tunnel-error metric where the client and gateway event-loops
already drain the errors, instead of in a dedicated `Tunnel::record_error`
pass. The error-chain encoding and counter constructor move to
`telemetry::otel` so both event-loops can share them.
