+++ rust/libs/connlib/tunnel/src/lib.rs
@@ -103,7 +103,6 @@ pub struct Tunnel<TRoleState> {
     buffers: Buffers,
 
     packet_counter: opentelemetry::metrics::Counter<u64>,
-    tunnel_errors: opentelemetry::metrics::Counter<u64>,
 }
 
 impl<TRoleState> Tunnel<TRoleState> {
@@ -117,12 +116,6 @@ impl<TRoleState> Tunnel<TRoleState> {
 
     pub fn rebind_dns(&mut self, sockets: Vec<SocketAddr>) -> Result<(), TunnelError> {
         self.io.rebind_dns(sockets)
-    }
-
-    fn record_error(&self, error: &TunnelError) {
-        for error in &error.errors {
-            self.tunnel_errors.add(1, &otel::error_layers(error));
-        }
     }
 }
 
@@ -153,7 +146,6 @@ impl ClientTunnel {
                 .u64_counter("system.network.packets")
                 .with_description("The number of packets processed.")
                 .build(),
-            tunnel_errors: otel::tunnel_errors(),
         }
     }
 
@@ -239,7 +231,7 @@ impl ClientTunnel {
             }
 
             // Process all IO sources that are ready.
-            let errors = if let Poll::Ready(io::Input {
+            if let Poll::Ready(io::Input {
                 now,
                 timeout,
                 dns_response,
@@ -324,15 +316,9 @@ impl ClientTunnel {
                     tick.want_continue();
                 }
 
-                error
-            } else {
-                TunnelError::default()
-            };
-
-            if !errors.is_empty() {
-                self.record_error(&errors);
-
-                return Poll::Ready(ClientEvent::Error(errors));
+                if !error.is_empty() {
+                    return Poll::Ready(ClientEvent::Error(error));
+                }
             }
         }
 
@@ -367,7 +353,6 @@ impl GatewayTunnel {
                 .u64_counter("system.network.packets")
                 .with_description("The number of packets processed.")
                 .build(),
-            tunnel_errors: otel::tunnel_errors(),
         }
     }
 
@@ -420,7 +405,7 @@ impl GatewayTunnel {
             }
 
             // Process all IO sources that are ready.
-            let errors = if let Poll::Ready(io::Input {
+            if let Poll::Ready(io::Input {
                 now,
                 timeout,
                 dns_response,
@@ -594,15 +579,9 @@ impl GatewayTunnel {
                     tick.want_continue();
                 }
 
-                error
-            } else {
-                TunnelError::default()
-            };
-
-            if !errors.is_empty() {
-                self.record_error(&errors);
-
-                return Poll::Ready(GatewayEvent::Error(errors));
+                if !error.is_empty() {
+                    return Poll::Ready(GatewayEvent::Error(error));
+                }
             }
         }
